requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==4.2.2
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
clip @ git+https://github.com/openai/clip.git@a1d071733d7111c9c014f024669f959182114e33
comm==0.1.4
ConfigArgParse==1.7
dataclasses-json==0.5.14
datasets==2.14.3
debugpy==1.6.8
decorator==5.1.1
dill==0.3.7
dnspython==2.4.1
entrypoints==0.4
executing==1.2.0
filelock==3.12.2
Flask==2.3.2
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
gevent==23.7.0
geventhttpclient==2.0.9
gitdb==4.0.10
GitPython==3.1.32
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpc-gateway-protoc-gen-openapiv2==0.1.0
grpcio==1.56.2
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
langchain==0.0.252
langsmith==0.0.18
locust==2.16.0
loguru==0.7.0
lz4==4.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.5.7
networkx==3.1
numexpr==2.8.4
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pinecone-client==2.2.2
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==3.19.6
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pydantic==1.10.12
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.30.1
regex==2023.6.3
requests==2.31.0
rich==13.5.2
roundrobin==0.0.4
rpds-py==0.9.2
safetensors==0.3.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.19
stack-data==0.6.2
streamlit==1.21.0
streamlit-chat==0.0.2.2
sympy==1.12
tenacity==8.2.2
tiktoken==0.3.3
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
uuid==1.30
validators==0.20.0
wcwidth==0.2.6
Werkzeug==2.3.6
xxhash==3.3.0
yarl==1.9.2
zipp==3.16.2
zope.event==5.0
zope.interface==6.0
